requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.22.3
torch==1.10.1+cu111
gym==0.12.5
tensorboardX==1.2
pyyaml==3.12
gitpython==3.0.8
protobuf==3.20.0
